+++ android/app/src/main/java/com/example/issue_tracker/common/LifecycleOwnerExtension.kt
@@ -1,0 +1,22 @@
+package com.example.issue_tracker.common
+
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.repeatOnLifecycle
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.CoroutineStart
+import kotlinx.coroutines.launch
+import kotlin.coroutines.CoroutineContext
+import kotlin.coroutines.EmptyCoroutineContext
+
+fun LifecycleOwner.repeatOnLifecycleExtension(
+    context: CoroutineContext = EmptyCoroutineContext,
+    start: CoroutineStart = CoroutineStart.DEFAULT,
+    state: Lifecycle.State = Lifecycle.State.STARTED,
+    block: suspend CoroutineScope.() -> Unit,
+) {
+    lifecycleScope.launch(context, start) {
+        lifecycle.repeatOnLifecycle(state, block)
+    }
+}
